% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{ga}
\alias{ga}
\title{A class to connstruct, execute and analyse a genetic algorithm.}
\description{
This class will store data during the execution of a genetic algorithm. As well as this,
it contains a number of methods to interact with the algorithm and UI functionality to
view the results and ammend parameters during execution.
}
\details{
To get started, using the new() method to begin.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{par}}{a named list containing function parameters}

\item{\code{dim}}{the dimensions of the population array}

\item{\code{population}}{the population array}

\item{\code{operators}}{an ordered list of the operator functions}

\item{\code{dependents}}{function used by the operators}

\item{\code{metrics}}{a list containing custom metrics stored during the execution of the simulation}

\item{\code{output_graphs}}{a list of functions that produce a 2D graph (UI only)}

\item{\code{time}}{the number of operators that have been executed}

\item{\code{paused}}{whether the simulation is paused or not (UI only)}

\item{\code{speed}}{the number of generations executed every second (UI only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ga$new()}}
\item \href{#method-add_operators}{\code{ga$add_operators()}}
\item \href{#method-add_operator}{\code{ga$add_operator()}}
\item \href{#method-add_metric}{\code{ga$add_metric()}}
\item \href{#method-add_dependents}{\code{ga$add_dependents()}}
\item \href{#method-add_dependent}{\code{ga$add_dependent()}}
\item \href{#method-next_operator}{\code{ga$next_operator()}}
\item \href{#method-next_iteration}{\code{ga$next_iteration()}}
\item \href{#method-generate_server}{\code{ga$generate_server()}}
\item \href{#method-add_output_graph}{\code{ga$add_output_graph()}}
\item \href{#method-add_output_graphs}{\code{ga$add_output_graphs()}}
\item \href{#method-run_ui}{\code{ga$run_ui()}}
\item \href{#method-save_data}{\code{ga$save_data()}}
\item \href{#method-clone}{\code{ga$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$new(dim, initial, parameters = list(), store_data = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_operators"></a>}}
\if{latex}{\out{\hypertarget{method-add_operators}{}}}
\subsection{Method \code{add_operators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_operators(operators)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_operator"></a>}}
\if{latex}{\out{\hypertarget{method-add_operator}{}}}
\subsection{Method \code{add_operator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_operator(operator)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metric"></a>}}
\if{latex}{\out{\hypertarget{method-add_metric}{}}}
\subsection{Method \code{add_metric()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_metric(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dependents"></a>}}
\if{latex}{\out{\hypertarget{method-add_dependents}{}}}
\subsection{Method \code{add_dependents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_dependents(dependents)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dependent"></a>}}
\if{latex}{\out{\hypertarget{method-add_dependent}{}}}
\subsection{Method \code{add_dependent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_dependent(dependent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-next_operator"></a>}}
\if{latex}{\out{\hypertarget{method-next_operator}{}}}
\subsection{Method \code{next_operator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$next_operator()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-next_iteration"></a>}}
\if{latex}{\out{\hypertarget{method-next_iteration}{}}}
\subsection{Method \code{next_iteration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$next_iteration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_server"></a>}}
\if{latex}{\out{\hypertarget{method-generate_server}{}}}
\subsection{Method \code{generate_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$generate_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_output_graph"></a>}}
\if{latex}{\out{\hypertarget{method-add_output_graph}{}}}
\subsection{Method \code{add_output_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_output_graph(parameters)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_output_graphs"></a>}}
\if{latex}{\out{\hypertarget{method-add_output_graphs}{}}}
\subsection{Method \code{add_output_graphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$add_output_graphs(parameters)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_ui"></a>}}
\if{latex}{\out{\hypertarget{method-run_ui}{}}}
\subsection{Method \code{run_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$run_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_data"></a>}}
\if{latex}{\out{\hypertarget{method-save_data}{}}}
\subsection{Method \code{save_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$save_data(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ga$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
